+++ Mods/S7_UI_Components_Library_b66d56c6-12f9-4abc-844f-0c30b89d32e4/Story/RawFiles/Lua/S7_UCL_Auxiliary.lua
@@ -1,0 +1,14 @@
+--  ===============================
+--  REGISTER PATH-OVERRIDE LISTENER
+--  ===============================
+
+Ext.RegisterListener(
+    "ModuleLoadStarted",
+
+    function()
+        Ext.AddPathOverride(
+            "Public/Game/GUI/msgBox.swf",
+            "Public/S7_UI_Components_Library_b66d56c6-12f9-4abc-844f-0c30b89d32e4/GUI/msgBox.swf"
+        )
+    end
+)